backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
av==13.1.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloup==3.0.8
dataclasses-json==0.6.7
decorator==5.2.1
filetype==1.2.0
frozenlist==1.7.0
glcontext==3.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isosurfaces==0.1.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
manim==0.19.0
manimpango==0.6.0
mapbox-earcut==1.0.3
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
multidict==6.6.4
mypy-extensions==1.1.0
networkx==3.5
numpy==2.3.2
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.28.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pyglet==2.1.8
pyglm==2.8.2
pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
scipy==1.16.1
screeninfo==0.8.1
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.43
srt==3.5.3
svgelements==1.9.6
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
fastapi
uvicorn
celery
redis
beanie
pymongo
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
bcrypt<4.0
supabase
pydantic[email]
fastapi-mail
